REF: misplaced DataFrame arithmetic tests (#33151)
